+++ src/test/theme.test.js
@@ -376,4 +376,15 @@ describe('theming (jsdom)', () => {
 
     expectCSSMatches(`.a { color: green; }`)
   })
+
+  // https://github.com/styled-components/styled-components/issues/596
+  it('should hoist static properties when using withTheme', () => {
+    class MyComponent extends React.Component {
+      static myStaticProperty: boolean = true;
+    }
+
+    const MyComponentWithTheme = withTheme(MyComponent);
+
+    expect(MyComponentWithTheme.myStaticProperty).toBe(true);
+  })
 })